+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.4] - 08/09/2019.
+
+* `OverlayContainer` made a `WidgetsBindingObserver` and the overlay is re-inserted on every call to `didChangeMetrics`.
+
+* `OverlayContainer` hidden when `dispose` is called.
+
 ## [0.0.2] - 24/04/2019.
 
 * Comments added for `OverlayContainer`.